requirements.txt
langchain==0.0.208
langchain-groq==0.1.0  
playwright==1.39.0
python-dotenv==1.0.0